refactor main.rb and add school_library_app.rb
